Added a rotation animation to the current pieces
I found a simple way to rotate images, so I added our first animation.
Could be a little nicer if it found the actual center of the image and
rotated it about that point. Right now I am just approximating the
rotation, but it looks pretty good.

The animation lasts 3 frames.